+++ app/src/main/AndroidManifest.xml
@@ -125,7 +125,7 @@
                 android:value=".views.activemeasurements.ActiveMeasurementsActivity" />
         </activity>
         <activity
-            android:name=".views.settings.CaptureCodeActivity"
+            android:name=".views.dashboard.CaptureCodeActivity"
             android:screenOrientation="portrait"
             tools:replace="screenOrientation" />
