+++ app/src/main/java/cl/niclabs/adkintunmobile/views/dashboard/MainActivity.java
@@ -1,26 +1,36 @@
 package cl.niclabs.adkintunmobile.views.dashboard;
 
+import android.Manifest;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.graphics.Point;
 import android.os.Bundle;
-import android.support.design.widget.NavigationView;
+import android.support.v4.app.ActivityCompat;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentTransaction;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.widget.DrawerLayout;
+import android.support.v4.content.ContextCompat;
 import android.support.v7.app.AppCompatActivity;
+import android.telephony.TelephonyManager;
+import android.util.Log;
 import android.view.Display;
-import android.view.MenuItem;
 import android.view.View;
+import android.widget.Toast;
 
 import com.crashlytics.android.Crashlytics;
 import com.github.amlcurran.showcaseview.ShowcaseView;
 import com.github.amlcurran.showcaseview.targets.PointTarget;
 import com.github.amlcurran.showcaseview.targets.Target;
 import com.github.amlcurran.showcaseview.targets.ViewTarget;
+import com.google.zxing.BarcodeFormat;
+import com.google.zxing.integration.android.IntentIntegrator;
+import com.google.zxing.integration.android.IntentResult;
+import com.loopj.android.http.AsyncHttpClient;
+import com.loopj.android.http.AsyncHttpResponseHandler;
+
+import org.json.JSONObject;
 
 import java.text.SimpleDateFormat;
 
@@ -33,23 +43,26 @@ import cl.niclabs.adkintunmobile.utils.display.DisplayDateManager;
 import cl.niclabs.adkintunmobile.utils.display.NotificationManager;
 import cl.niclabs.adkintunmobile.utils.display.ShowCaseTutorial;
 import cl.niclabs.adkintunmobile.utils.information.Network;
-import cl.niclabs.adkintunmobile.views.aboutus.AboutUsActivity;
 import cl.niclabs.adkintunmobile.views.activeconnections.ActiveConnectionsActivity;
 import cl.niclabs.adkintunmobile.views.activemeasurements.ActiveMeasurementsActivity;
 import cl.niclabs.adkintunmobile.views.applicationstraffic.ApplicationsTrafficActivity;
 import cl.niclabs.adkintunmobile.views.connectiontype.connectionmode.ConnectionModeActivity;
 import cl.niclabs.adkintunmobile.views.connectiontype.networktype.NetworkTypeActivity;
 import cl.niclabs.adkintunmobile.views.notificationlog.NotificationLogActivity;
-import cl.niclabs.adkintunmobile.views.rankings.RankingFragment;
 import cl.niclabs.adkintunmobile.views.settings.SettingsActivity;
 import cl.niclabs.adkintunmobile.views.status.DataQuotaDialog;
 import cl.niclabs.adkintunmobile.views.status.DayOfRechargeDialog;
 import cl.niclabs.adkintunmobile.views.status.StatusActivity;
+import cz.msebera.android.httpclient.Header;
+import cz.msebera.android.httpclient.entity.StringEntity;
 import io.fabric.sdk.android.Fabric;
 
 public class MainActivity extends AppCompatActivity {
 
     private Context context;
+    private static final int REQUEST_READ_PHONE_STATE = 1;
+    static final String ACTION_SCAN = "com.google.zxing.client.android.SCAN";
+    private final String TAG = "AdkM:MainActivity";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -119,6 +132,10 @@ public class MainActivity extends AppCompatActivity {
         startActivity(myIntent);
     }
 
+    public void openAdkintunWebView(View view) {
+        checkReadPhonePermission();
+    }
+
     /*
      * changeCurrentFragment Methods
      */
@@ -129,6 +146,83 @@ public class MainActivity extends AppCompatActivity {
 
         fragmentTransaction.replace(R.id.main_content, newFragment, "DashboardFragment");
         fragmentTransaction.commit();
+    }
+
+    private void checkReadPhonePermission() {
+        final IntentIntegrator scanIntegrator = new IntentIntegrator(this);
+
+        int permissionCheck = ContextCompat.checkSelfPermission(this, Manifest.permission.READ_PHONE_STATE);
+        if (permissionCheck != PackageManager.PERMISSION_GRANTED) {
+            ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.READ_PHONE_STATE}, REQUEST_READ_PHONE_STATE);
+        } else {
+            scanIntegrator.setPrompt("");
+            scanIntegrator.setBeepEnabled(false);
+            scanIntegrator.setCaptureActivity(CaptureCodeActivity.class);
+            scanIntegrator.initiateScan();
+        }
+    }
+
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+
+        if(data != null && data.getAction() != null && data.getAction().equals(ACTION_SCAN)) {
+
+            IntentResult scanningResult = IntentIntegrator.parseActivityResult(requestCode, resultCode, data);
+            if (scanningResult != null) {
+                String qrString = scanningResult.getContents();
+                String scanFormat = scanningResult.getFormatName();
+
+                String qrCodeFormat = BarcodeFormat.QR_CODE.toString();
+                if (scanFormat.equals(qrCodeFormat)) {
+                    TelephonyManager telephonyManager = (TelephonyManager)getSystemService(Context.TELEPHONY_SERVICE);
+                    String deviceId = telephonyManager.getDeviceId();
+                    sendToServer(deviceId, qrString);
+                    showText(getString(R.string.settings_adkintun_web_qr_scanner_success));
+                }
+                else
+                    showText(getString(R.string.settings_adkintun_web_qr_scanner_failure));
+
+                super.onActivityResult(requestCode, resultCode, data);
+            }
+            else {
+                showText(getString(R.string.settings_adkintun_web_qr_scanner_failure));
+            }
+
+        }
+    }
+
+    private void sendToServer(String accessToken,String qrString) {
+        String url = getString(R.string.web_auth_url);
+
+        JSONObject params = new JSONObject();
+        try {
+            params.put("uuid", qrString);
+            params.put("access_token", accessToken);
+            StringEntity entity = new StringEntity(params.toString());
+
+            AsyncHttpClient client = new AsyncHttpClient();
+            client.post(this, url, entity, "application/json", new AsyncHttpResponseHandler() {
+                @Override
+                public void onSuccess(int statusCode, Header[] headers, byte[] responseBody) {
+                    Log.d(TAG, statusCode + " OK");
+                }
+
+                @Override
+                public void onFailure(int statusCode, Header[] headers, byte[] responseBody, Throwable error) {
+                    Log.d(TAG, statusCode + " NOT OK");
+                }
+            });
+
+        } catch (Exception e) {
+            e.printStackTrace();
+            return;
+        }
+    }
+
+    private void showText(String message){
+        Toast toast = Toast.makeText(getApplicationContext(),
+                message, Toast.LENGTH_SHORT);
+        toast.show();
     }
 
     public void notif(View view){
