+++ app/src/main/java/cl/niclabs/adkintunmobile/views/settings/SettingsActivity.java
@@ -27,7 +27,6 @@ public class SettingsActivity extends AppCompatActivity {
 
     private Toolbar toolbar;
     private Context context;
-    static final String ACTION_SCAN = "com.google.zxing.client.android.SCAN";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -53,68 +52,5 @@ public class SettingsActivity extends AppCompatActivity {
         getSupportActionBar().setDisplayShowHomeEnabled(true);
 
         setTitle(getString(R.string.view_settings));
-    }
-
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-
-        if(data != null && data.getAction() != null && data.getAction().equals(ACTION_SCAN)) {
-
-            IntentResult scanningResult = IntentIntegrator.parseActivityResult(requestCode, resultCode, data);
-            if (scanningResult != null) {
-                String qrString = scanningResult.getContents();
-                String scanFormat = scanningResult.getFormatName();
-
-                String qrCodeFormat = BarcodeFormat.QR_CODE.toString();
-                if (scanFormat.equals(qrCodeFormat)) {
-                    TelephonyManager telephonyManager = (TelephonyManager)getSystemService(Context.TELEPHONY_SERVICE);
-                    String deviceId = telephonyManager.getDeviceId();
-                    sendToServer(deviceId, qrString);
-                    showText(getString(R.string.settings_adkintun_web_qr_scanner_success));
-                }
-                else
-                    showText(getString(R.string.settings_adkintun_web_qr_scanner_failure));
-
-                super.onActivityResult(requestCode, resultCode, data);
-            }
-            else {
-                showText(getString(R.string.settings_adkintun_web_qr_scanner_failure));
-            }
-
-        }
-    }
-
-    private void sendToServer(String accessToken,String qrString) {
-        String url = getString(R.string.web_auth_url);
-
-        JSONObject params = new JSONObject();
-        try {
-            params.put("uuid", qrString);
-            params.put("access_token", accessToken);
-            StringEntity entity = new StringEntity(params.toString());
-
-            AsyncHttpClient client = new AsyncHttpClient();
-            client.post(this, url, entity, "application/json", new AsyncHttpResponseHandler() {
-                @Override
-                public void onSuccess(int statusCode, Header[] headers, byte[] responseBody) {
-                    Log.d(TAG, statusCode + " OK");
-                }
-
-                @Override
-                public void onFailure(int statusCode, Header[] headers, byte[] responseBody, Throwable error) {
-                    Log.d(TAG, statusCode + " NOT OK");
-                }
-            });
-
-        } catch (Exception e) {
-            e.printStackTrace();
-            return;
-        }
-    }
-
-    private void showText(String message){
-        Toast toast = Toast.makeText(getApplicationContext(),
-                message, Toast.LENGTH_SHORT);
-        toast.show();
     }
 }
