+++ app/src/main/res/layout/activity_capture_code.xml
@@ -5,7 +5,7 @@
     android:orientation="vertical"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".views.settings.CaptureCodeActivity">>
+    tools:context=".views.dashboard.CaptureCodeActivity">>
 
 
     <!-- Toolbar -->
