+++ app/src/main/res/layout/fragment_dashboard.xml
@@ -176,7 +176,7 @@
                 android:layout_height="match_parent">
 
                 <android.support.v7.widget.CardView
-                    android:layout_width="match_parent"
+                    android:layout_width="0dp"
                     android:layout_height="match_parent"
                     android:layout_marginEnd="4dp"
                     android:layout_marginRight="4dp"
@@ -216,7 +216,7 @@
                 </android.support.v7.widget.CardView>
 
                 <android.support.v7.widget.CardView
-                    android:layout_width="match_parent"
+                    android:layout_width="0dp"
                     android:layout_height="match_parent"
                     android:layout_marginStart="4dp"
                     android:layout_marginLeft="4dp"
@@ -367,49 +367,96 @@
 
         </android.support.v7.widget.CardView>
 
-        <android.support.v7.widget.CardView
+
+        <TableLayout
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
+            android:layout_height="match_parent"
             android:layout_marginTop="@dimen/separation_little"
-            android:id="@+id/card_preferences"
-            android:visibility="visible">
+            android:stretchColumns="0,1">
 
-            <LinearLayout
-                style="@style/Widget.CardContent"
+            <TableRow
                 android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:onClick="openPreferencesView">
+                android:layout_height="match_parent">
 
-                <TextView
-                    android:layout_width="match_parent"
-                    android:layout_height="wrap_content"
-                    android:text="@string/view_dashboard_preferences_title"
-                    android:textAppearance="@style/TextAppearance.AppCompat.Title" />
-                <TableLayout
-                    android:layout_width="match_parent"
+                <android.support.v7.widget.CardView
+                    android:layout_width="0dp"
                     android:layout_height="match_parent"
-                    android:stretchColumns="*">
+                    android:layout_marginEnd="4dp"
+                    android:layout_marginRight="4dp"
+                    android:id="@+id/card_preferences"
+                    android:visibility="visible">
 
-                    <TableRow>
+                    <LinearLayout
+                        style="@style/Widget.CardContent"
+                        android:layout_width="match_parent"
+                        android:layout_height="match_parent"
+                        android:onClick="openPreferencesView">
+
                         <TextView
-                            android:layout_width="0dp"
+                            android:layout_width="match_parent"
                             android:layout_height="wrap_content"
-                            android:layout_weight="4"
+                            android:text="@string/view_dashboard_preferences_title"
+                            android:textAppearance="@style/TextAppearance.AppCompat.Title" />
+
+                        <TextView
+                            android:layout_width="match_parent"
+                            android:layout_height="wrap_content"
                             android:text="@string/view_dashboard_preferences_summary" />
 
-                        <ImageView
-                            android:layout_width="0dp"
+                        <LinearLayout
+                            android:layout_width="match_parent"
+                            android:layout_height="match_parent">
+                            <ImageView
+                                android:layout_width="match_parent"
+                                android:layout_height="wrap_content"
+                                android:layout_gravity="center_vertical"
+                                android:id="@+id/iv_preferences"
+                                android:alpha="0.6"
+                                android:src="@drawable/ic_settings_black"/>
+                        </LinearLayout>
+
+
+                    </LinearLayout>
+
+                </android.support.v7.widget.CardView>
+
+                <android.support.v7.widget.CardView
+                    android:layout_width="0dp"
+                    android:layout_height="match_parent"
+                    android:layout_marginStart="4dp"
+                    android:layout_marginLeft="4dp"
+                    android:id="@+id/card_adkintun_web_scan"
+                    android:visibility="visible">
+
+                    <LinearLayout
+                        style="@style/Widget.CardContent"
+                        android:layout_width="match_parent"
+                        android:layout_height="match_parent"
+                        android:onClick="openAdkintunWebView">
+
+                        <TextView
+                            android:layout_width="match_parent"
                             android:layout_height="wrap_content"
-                            android:id="@+id/iv_preferences"
-                            android:layout_weight="1"
+                            android:text="@string/view_dashboard_adkintun_web_title"
+                            android:textAppearance="@style/TextAppearance.AppCompat.Title" />
+
+                        <TextView
+                            android:layout_width="match_parent"
+                            android:layout_height="wrap_content"
+                            android:text="@string/view_dashboard_adkintun_web_summary" />
+
+                        <ImageView
+                            android:layout_width="match_parent"
+                            android:layout_height="wrap_content"
+                            android:layout_marginTop="@dimen/separation_little"
+                            android:id="@+id/iv_adkintun_web"
                             android:alpha="0.6"
-                            android:src="@drawable/ic_settings_black"/>
-                    </TableRow>
-                </TableLayout>
+                            android:src="@drawable/ic_devices_black"/>
+                    </LinearLayout>
 
-            </LinearLayout>
-
-        </android.support.v7.widget.CardView>
+                </android.support.v7.widget.CardView>
+            </TableRow>
+        </TableLayout>
 
     </LinearLayout>
