+++ app/src/main/res/values/strings.xml
@@ -30,9 +30,6 @@
     <string name="settings_app_daily_notification_title">Notificaciones Diarias</string>
     <string name="settings_app_daily_notification_summary">Programar una notificación diaria para el final del día con el resumen del consumo de datos móviles.</string>
 
-    <string name="settings_adkintun_web_category">Adkintun Web</string>
-    <string name="settings_adkintun_web_qr_scanner_title">Escanear Código QR</string>
-    <string name="settings_adkintun_web_qr_scanner_key">adkintun_qr_scanner</string>
     <string name="settings_adkintun_web_qr_scanner_prompt">Ingresa a web.adkintunmobile.cl y escanea el código QR para acceder a tu información.</string>
     <string name="settings_adkintun_web_qr_scanner_success">Código QR escaneado</string>
     <string name="settings_adkintun_web_qr_scanner_failure">Error al escanear código</string>
@@ -131,6 +128,8 @@
     <string name="view_dashboard_active_connections_summary">En este instante las conexiones de internet activas en tu celular llegan a un total de:</string>
     <string name="view_dashboard_preferences_title">Preferencias</string>
     <string name="view_dashboard_preferences_summary">Edita las preferencias de Adkintun Mobile</string>
+    <string name="view_dashboard_adkintun_web_title">Adkintun Mobile Web</string>
+    <string name="view_dashboard_adkintun_web_summary">Escanea el código QR y accede a tu información en adkintunmobile.cl</string>
 
 
     <!-- Notifications Log -->
